Refactoring around claims export service with the aim of making it a general purpose export service
